// src/lib/enums/transactions/transaction-internal-status.enum.ts
export enum TransactionInternalStatus {
  // status: NOT_STARTED
  UNKNOWN = 'UNKNOWN',

  // DEPOSIT
  DEPOSIT_CREATED = 'DEPOSIT_CREATED', // status: PENDING

  DEPOSIT_SUBMITTED_TO_UPSTREAM = 'DEPOSIT_SUBMITTED_TO_UPSTREAM', // status: PENDING
  DEPOSIT_SUBMIT_TO_UPSTREAM_ERROR = 'DEPOSIT_SUBMIT_TO_UPSTREAM_ERROR', // status: FAILED
  DEPOSIT_UPSTREAM_CONNECTION_ERROR = 'DEPOSIT_UPSTREAM_CONNECTION_ERROR', // status: FAILED

  DEPOSIT_UPSTREAM_SUCCESS_RESPONSE = 'DEPOSIT_UPSTREAM_SUCCESS_RESPONSE', // status: SUCCESS
  DEPOSIT_UPSTREAM_FAILED_RESPONSE = 'DEPOSIT_UPSTREAM_FAILED_RESPONSE', // status: FAILED

  // WITHDRAWAL
  WITHDRAWAL_CREATED = 'WITHDRAWAL_CREATED', // status: PENDING

  WITHDRAWAL_SUBMITTED_TO_UPSTREAM = 'WITHDRAWAL_SUBMITTED_TO_UPSTREAM', // status: PENDING
  WITHDRAWAL_SUBMIT_TO_UPSTREAM_ERROR = 'WITHDRAWAL_SUBMIT_TO_UPSTREAM_ERROR', // status: FAILED
  WITHDRAWAL_UPSTREAM_CONNECTION_ERROR = 'WITHDRAWAL_UPSTREAM_CONNECTION_ERROR', // status: FAILED

  WITHDRAWAL_UPSTREAM_INSUFFICIENT_BALANCE = 'WITHDRAWAL_UPSTREAM_INSUFFICIENT_BALANCE', // status: PENDING (Problem Transaction)
  WITHDRAWAL_UPSTREAM_SUCCESS_RESPONSE = 'WITHDRAWAL_UPSTREAM_SUCCESS_RESPONSE', // status: SUCCESS
  WITHDRAWAL_UPSTREAM_FAILED_RESPONSE = 'WITHDRAWAL_UPSTREAM_FAILED_RESPONSE', // status: FAILED

  WITHDRAWAL_UPSTREAM_FAILED_REFUNDED_RESPONSE = 'WITHDRAWAL_UPSTREAM_FAILED_REFUNDED_RESPONSE', // status: FAILED_WITHDRAWAL_REFUNDED
  WITHDRAWAL_REFUNDED_TO_MERCHANT = 'WITHDRAWAL_REFUNDED_TO_MERCHANT', // status: FAILED_WITHDRAWAL_REFUNDED (revert)

  // MERCHANT_REQUESTED_WITHDRAWAL
  MERCHANT_REQUESTED_WITHDRAWAL_CREATED = 'MERCHANT_REQUESTED_WITHDRAWAL_CREATED', // status: PENDING
  MERCHANT_REQUESTED_WITHDRAWAL_REJECTED = 'MERCHANT_REQUESTED_WITHDRAWAL_REJECTED', // status: FAILED_MERCHANT_REQUESTED_WITHDRAWAL_REJECTED
}