// src/lib/enums/transactions/transaction-status.enum.ts
export enum TransactionStatus {
  // Intermediate statuses
  NOT_STARTED = 'NOT_STARTED',
  PENDING = 'PENDING',
  MERCHANT_REQUESTED_WITHDRAWAL_PENDING = 'MERCHANT_REQUESTED_WITHDRAWAL_PENDING',

  // Final statuses
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
  FAILED_WITHDRAWAL_REFUNDED = 'FAILED_WITHDRAWAL_REFUNDED', // for refunds in API_WITHDRAWAL or MERCHANT_REQUESTED_WITHDRAWAL
  FAILED_MERCHANT_REQUESTED_WITHDRAWAL_REJECTED = 'FAILED_MERCHANT_REQUESTED_WITHDRAWAL_REJECTED',
}

/**
 * Possible status transitions:
 * API_DEPOSIT:
 *   NOT_STARTED -> PENDING -> SUCCESS
 *   NOT_STARTED -> PENDING -> FAILED
 * API_WITHDRAWAL:
 *   NOT_STARTED -> PENDING -> SUCCESS
 *   NOT_STARTED -> PENDING -> SUCCESS -> FAILED_WITHDRAWAL_REFUNDED
 *   NOT_STARTED -> PENDING -> FAILED
 * MERCHANT_REQUESTED_WITHDRAWAL:
 *   MERCHANT_REQUESTED_WITHDRAWAL_PENDING -> FAILED_MERCHANT_REQUESTED_WITHDRAWAL_REJECTED
 *   MERCHANT_REQUESTED_WITHDRAWAL_PENDING -> PENDING -> SUCCESS
 *   MERCHANT_REQUESTED_WITHDRAWAL_PENDING -> PENDING -> SUCCESS -> FAILED_WITHDRAWAL_REFUNDED
 *   MERCHANT_REQUESTED_WITHDRAWAL_PENDING -> PENDING -> FAILED
 */